[SWM-185] Design : MainDashboard - WeeklyCalendar
